+++ profile_base/__manifest__.py
@@ -6,7 +6,7 @@
 #
 {
     "name": "Base profile",
-    "version": "10.0.1.0.3",
+    "version": "10.0.1.0.4",
     "category": "Technical Settings",
     "summary": "Base setup for minimal configuration Zeroincombenze(R)",
     "author": "SHS-AV s.r.l.",
@@ -15,14 +15,23 @@
     "license": "AGPL-3",
     "depends": [
         "base",
+        # "remove_odoo_enterprise",           # OCA server-brand - Only 12.0
+        # "disable_odoo_online",              # OCA server-brand - Only 12.0
         "l10n_it_coa",
-        "partner_bank",
+        "partner_bank",                         # Only 10.0
+        # "portal_odoo_debranding",           # OCA server-brand - Only 12.0
+        # "partner_data_from_vies",
         "purchase",
         "purchase_discount",
         "report_xlsx",
+        # "repository_check",                  # custom-addons - Only 12.0
         "sale",
+        # "sale_management",  # Only 12.0
         "stock",
         "web_sheet_full_width",
+        "product_by_supplier",      # only 10.0
+        "sale_order_dates",         # Only 10.0
+        "delivery",
     ],
     "maintainer": "Antonio M. Vigliotti <antoniomaria.vigliotti@gmail.com>",
     "installable": True,